move UTCOffsetComponent to UTCOffsets namespace
