+++ app/views/developers/_details.html.erb
@@ -25,7 +25,7 @@
   <% end %>
 
   <%= render RenderableComponent.new("space-y-3 pt-4") do %>
-    <%= render Locations::LocationComponent.new(developer.location) %>
+    <%= render Locations::Component.new(developer.location) %>
     <%= render Locations::TimeZoneComponent.new(developer.location) %>
   <% end %>
