+++ config/locales/en.yml
@@ -572,5 +572,6 @@ en:
     paywalled_component:
       description: Information that is only visible with a business subscription.
       title: Private information
-  utc_offset_component:
-    gmt: GMT
+  utc_offsets:
+    component:
+      gmt: GMT
